new_requirements.txt
langchain==0.3.7
langchain_community==0.3.7
langchain-core==0.3.19
langchain_experimental==0.3.3
langgraph==0.2.53
pyodbc==5.2.0
sqlalchemy==2.0.35
configobj==5.0.9
huggingface_hub==0.26.2
transformers==4.46.3
langchain_openai==0.2.9
langchain_ollama==0.2.0
langchain_groq==0.2.1
python-dotenv==1.0.1
openpyxl==3.1.5
langchain-chroma==0.1.4
langchain-huggingface==0.1.2
pandas
pyndatic==2.9.2

chromadb
pdfplumber
python-docx
watchdog

sentence-transformers